refactor(ConvertRawLecturesUseCase):add mock files
